requirement.txt
Appium-Python-Client==1.0.2
selenium==3.141.0

